feat (statistics) : 가게 대시보드 구현

+++ src/main/java/com/example/outsourcing/domain/statistics/dto/response/StatisticsPriceResponseDto.java
@@ -9,5 +9,5 @@ public class StatisticsPriceResponseDto {
 
     private final String storeName;
 
-    private final Integer totalPrice;
+    private final Long totalPrice;
 }
